+++ src/Geo/Resources/RegionResource.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Enraiged\Geo\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class RegionResource extends JsonResource
+{
+    /**
+     *  Transform the resource collection into an array.
+     *
+     *  @param  \Illuminate\Http\Request  $request
+     *  @return array
+     */
+    public function toArray($request): array
+    {
+        return [
+            'id' => $this->id,
+            'code' => $this->code,
+            'name' => $this->name,
+            'type' => $this->type,
+            'is_active' => $this->is_active,
+        ];
+    }
+}